lib: initial aarch64 support
This change imports Hayashi Naoyuki's Aarch64 work into lib

+++ usr/src/lib/krb5/plugins/preauth/pkinit/pkinit_crypto_openssl.c
@@ -34,6 +34,9 @@
  * Copyright 2018 OmniOS Community Edition (OmniOSce) Association.
  * Copyright 2018 RackTop Systems.
  */
+/*
+ * Copyright 2017 Hayashi Naoyuki
+ */
 
 #include <errno.h>
 #include <string.h>
@@ -4713,7 +4716,7 @@ pkinit_C_Decrypt(pkinit_identity_crypto_context id_cryptoctx,
     rv = id_cryptoctx->p11->C_Decrypt(id_cryptoctx->session, pEncryptedData,
 	ulEncryptedDataLen, pData, pulDataLen);
     if (rv == CKR_OK) {
-	pkiDebug("pData %x *pulDataLen %d\n", (int) pData, (int) *pulDataLen);
+	pkiDebug("pData %x *pulDataLen %d\n", (int) (intptr_t)pData, (int) *pulDataLen);
     }
     return rv;
 }
@@ -4772,8 +4775,8 @@ pkinit_decode_data_pkcs11(krb5_context context,
     len = data_len;
 #ifdef SILLYDECRYPT
     pkiDebug("session %x edata %x edata_len %d data %x datalen @%x %d\n",
-	    (int) id_cryptoctx->session, (int) data, (int) data_len, (int) cp,
-	    (int) &len, (int) len);
+	    (int) id_cryptoctx->session, (int)(intptr_t)data, (int) data_len, (int) (intptr_t) cp,
+	    (int)  (intptr_t)&len, (int) len);
     if ((r = pkinit_C_Decrypt(id_cryptoctx, data, (CK_ULONG) data_len,
 	    cp, &len)) != CKR_OK) {
 #else
